core: Move a few modules to ./utils/
- Makes clear they are not related to the synchronization domain
- Makes it easier to see the main synchronization modules

Moved modules:

- logger
- perftools (renamed to utils/perfs)
- sentry
- timestamp
