+++ core/local/atom/add_checksum.js
@@ -3,7 +3,7 @@
 const _ = require('lodash')
 const path = require('path')
 
-const logger = require('../../logger')
+const logger = require('../../utils/logger')
 
 const STEP_NAME = 'addChecksum'
