+++ core/local/atom/overwrite.js
@@ -3,7 +3,7 @@
 const _ = require('lodash')
 
 const Channel = require('./channel')
-const logger = require('../../logger')
+const logger = require('../../utils/logger')
 
 /*::
 import type { AtomEvent, AtomBatch } from './event'
