+++ core/local/checksumer.js
@@ -4,7 +4,8 @@ const Promise = require('bluebird')
 const async = require('async')
 const crypto = require('crypto')
 const fs = require('fs')
-const measureTime = require('../perftools')
+
+const measureTime = require('../utils/perfs')
 
 /*::
 import type { Callback } from '../utils/func'
