+++ core/remote/cozy.js
@@ -12,7 +12,7 @@ const {
   keepFiles,
   parentDirIds
 } = require('./document')
-const logger = require('../logger')
+const logger = require('../utils/logger')
 
 const { posix } = path
