+++ core/utils/fs.js
@@ -1,7 +1,7 @@
 const Promise = require('bluebird')
 const childProcess = require('child_process')
 
-const logger = require('../logger')
+const logger = require('./logger')
 
 Promise.promisifyAll(childProcess)
