+++ core/utils/perfs.js
